release: 0.0.1-beta.4 (#439)
* [429] type error triage (#435)

* refactor: make builder a typescript application

* fix: missing font in dark theme type

* update: dark theme

* Fix decimals for mainnet ETH token (#438)

* release: 0.0.1-beta.4 (#440)

---------
