+++ wormhole-connect-loader/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@wormhole-foundation/wormhole-connect",
-  "version": "0.0.1-beta.3",
+  "version": "0.0.1-beta.4",
   "dependencies": {
     "@mui/material": "^5.12.1"
   },
